Use unsigned long instead of unsigned for the integer casts here.  The former works for both ILP32 and LP64 programming models, while the latter fails LP64.
# uintpr_t is better, but iirc, we can't pollute the name space to use it
# I likely need to audit all my uintptr_t changes for that issue...


Former-commit-id: 2b23d9afd06bc1900edf8e309d419f06266eef39

+++ sys/mips/include/param.h
@@ -491,7 +491,7 @@ name|mips_round_page
 parameter_list|(
 name|x
 parameter_list|)
-value|((((unsigned)(x)) + NBPG - 1)& ~(NBPG-1))
+value|((((unsigned long)(x)) + NBPG - 1)& ~(NBPG-1))
 end_define
 
 begin_define
@@ -501,7 +501,7 @@ name|mips_trunc_page
 parameter_list|(
 name|x
 parameter_list|)
-value|((unsigned)(x)& ~(NBPG-1))
+value|((unsigned long)(x)& ~(NBPG-1))
 end_define
 
 begin_define
@@ -511,7 +511,7 @@ name|mips_btop
 parameter_list|(
 name|x
 parameter_list|)
-value|((unsigned)(x)>> PGSHIFT)
+value|((unsigned long)(x)>> PGSHIFT)
 end_define
 
 begin_define
@@ -521,7 +521,7 @@ name|mips_ptob
 parameter_list|(
 name|x
 parameter_list|)
-value|((unsigned)(x)<< PGSHIFT)
+value|((unsigned long)(x)<< PGSHIFT)
 end_define
 
 begin_define
